+++ module/helpers/config.mjs
@@ -5,26 +5,138 @@ export const slaindustries = {};
  * @type {Object}
  */
  slaindustries.abilities = {
-  "str": "slaindustries.AbilityStr",
-  "dex": "slaindustries.AbilityDex",
-  "know": "slaindustries.AbilityKnow",
-  "conc": "slaindustries.AbilityConc",
-  "cha": "slaindustries.AbilityCha",
-  "cool": "slaindustries.AbilityCool",
-  "luck": "slaindustries.AbilityLuck",
-  "closing": "slaindustries.AbilityClosing",
-  "rush": "slaindustries.AbilityRush"
+  str: "slaindustries.abilities.str",
+  dex: "slaindustries.abilities.dex",
+  know: "slaindustries.abilities.know",
+  conc: "slaindustries.abilities.conc",
+  cha: "slaindustries.abilities.cha",
+  cool: "slaindustries.abilities.cool",
 };
 
 slaindustries.abilityAbbreviations = {
-  "str": "slaindustries.AbilityStrAbbr",
-  "dex": "slaindustries.AbilityDexAbbr",
-  "know": "slaindustries.AbilityKnowAbbr",
-  "conc": "slaindustries.AbilityConcAbbr",
-  "cha": "slaindustries.AbilityChaAbbr",
-  "cool": "slaindustries.AbilityCoolAbbr",
-  "luck": "slaindustries.AbilityLuckAbbr",
-  "closing": "slaindustries.ClosingChaAbbr",
-  "rush": "slaindustries.AbilityRushAbbr"
-  
+  str: "slaindustries.abilityAbbr.str",
+  dex: "slaindustries.abilityAbbr.dex",
+  know: "slaindustries.abilityAbbr.know",
+  conc: "slaindustries.abilityAbbr.conc",
+  cha: "slaindustries.abilityAbbr.cha",
+  cool: "slaindustries.abilityAbbr.cool",
+  luck: "slaindustries.abilityAbbr.luck",
+};
+
+slaindustries.rangeSkills = {
+    pistol: "slaindustries.skills.pistol",
+    rifle: "slaindustries.skills.rifle",
+    support: "slaindustries.skills.support",
+};
+
+slaindustries.meleeSkills = {
+    melee: "slaindustries.skills.melee",
+    unarmed: "slaindustries.skills.unarmed",
+    pole: "slaindustries.skills.pole",
+    shield: "slaindustries.skills.shield",
+};
+
+slaindustries.strSkills = {
+    climb: "slaindustries.skills.climb",
+    melee: "slaindustries.skills.melee",
+    pole: "slaindustries.skills.pole",
+    shield: "slaindustries.skills.shield",
+    support: "slaindustries.skills.support",
+    swim: "slaindustries.skills.swim",
+    throw: "slaindustries.skills.throw",
+    unarmed:"slaindustries.skills.unarmed",
+};
+
+slaindustries.dexSkills = {
+    acro: "slaindustries.skills.acro",
+    ath: "slaindustries.skills.ath",
+    moto: "slaindustries.skills.moto",
+    forge: "slaindustries.skills.forge",
+    pistol: "slaindustries.skills.pistol",
+    rifle: "slaindustries.skills.rifle",
+    sleight: "slaindustries.skills.sleight",
+    stealth: "slaindustries.skills.stealth",
+};
+
+slaindustries.knowSkills = {
+    admin: "slaindustries.skills.admin",
+    comp: "slaindustries.skills.comp",
+    edu: "slaindustries.skills.edu",
+    forensics: "slaindustries.skills.forensics",
+    lang: "slaindustries.skills.lang",
+    lore: "slaindustries.skills.lore",
+    med: "slaindustries.skills.med",
+    street: "slaindustries.skills.street",
+};
+
+slaindustries.concSkills = {
+    demo: "slaindustries.skills.demo",
+    drive: "slaindustries.skills.drive",
+    detect: "slaindustries.skills.detect",
+    pick: "slaindustries.skills.pick",
+    lips: "slaindustries.skills.lips",
+    tac: "slaindustries.skills.tac",
+    tech: "slaindustries.skills.tech",
+    track: "slaindustries.skills.track",
+};
+
+slaindustries.chaSkills = {
+    dip: "slaindustries.skills.dip",
+    haggle: "slaindustries.skills.haggle",
+    lead: "slaindustries.skills.lead",
+    oratory: "slaindustries.skills.oratory",
+    per: "slaindustries.skills.per",
+    sed: "slaindustries.skills.sed",
+};
+
+slaindustries.coolSkills = {
+    bribe: "slaindustries.skills.bribe",
+    gamble: "slaindustries.skills.gamble",
+    intero: "slaindustries.skills.intero",
+    intimid: "slaindustries.skills.intimid",
+    surv: "slaindustries.skills.surv",
+    tort: "slaindustries.skills.tort",
+};
+
+slaindustries.ebb = {
+    aware: "slaindustries.ebb.aware",
+    blast: "slaundustries.ebb.blast",
+    comm: "slaindustries.ebb.comm",
+    enh: "slaindustries.ebb.enh",
+    heal: "slaindustries.ebb.heal",
+    prot: "slaindustries.ebb.prot",
+    fold: "slaindustries.ebb.fold",
+    sense: "slaindustries.ebb.sense",
+    tele: "slaindustries.ebb.tele",
+    therm: "slaindustries.ebb.therm",
+};
+
+slaindustries.species = {
+    human: "slaindustries.species.human",
+    froth: "slaindustries.species.froth",
+    ebb: "slaindustries.species.ebb",
+    storm_313: "slaindustries.species.storm_313",
+    storm_717: "slaindustries.species.storm_717",
+    shaktar: "slaindustries.species.shaktar",
+    wraith: "slaindustries.species.wraith",
+    advCar: "slaindustries.species.advCar",
+    neo: "slaindustries.species.neo",
+};
+
+slaindustries.package = {
+    basic: "slaindustries.package.basic",
+    advanced: "slaindustries.package.advanced",
+    elite: "slaindustries.package.elite",
+    master: "slaindustries.package.master",
+};
+
+slaindustries.package = {
+    strike: "slaindustries.package.strike",
+    cqc: "slaindustries.package.cqc",
+    heavy: "slaindustries.package.heavy",
+    scout: "slaindustries.package.scout",
+    medic: "slaindustries.package.medic",
+    investigation: "slaindustries.package.investigation",
+    tech: "slaindustries.package.tech",
+    beauraucrat: "slaindustries.package.beauraucrat",
 };