+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
-## 1.2.0
-
-- Add support for Foundry v10
+## 0.4.35	
+	- Removed levels in the data model attributes should now be system.abilities.KEY (system.abilities.cool)
+	- Improved skill selector dropdowns for melee and ranged weapons
+	- Added a dropdown for species and package.
+	- Added drug items
+	- Made minor layout fixes
+	- Added better support for localization